pkg-py/examples/ibis-mtcars/requirements.txt
querychat[ibis]
pandas
sqlalchemy
ibis-sqlite
plotly
shinywidgets